refactor: simplify mutations and extract meter creation hook
