+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/services/-services/service-benefits-tab.tsx
@@ -11,11 +11,7 @@ import {
    PopoverContent,
    PopoverTrigger,
 } from "@packages/ui/components/popover";
-import {
-   useMutation,
-   useQueryClient,
-   useSuspenseQueries,
-} from "@tanstack/react-query";
+import { useMutation, useSuspenseQueries } from "@tanstack/react-query";
 import {
    formatCostBRL,
    summarizeByType,
@@ -34,6 +30,7 @@ import { orpc } from "@/integrations/orpc/client";
 import { BenefitAttachPopover } from "./benefit-attach-popover";
 import { ServiceMarginPanel } from "./service-margin-panel";
 import { ServiceTabToolbar } from "./service-tab-toolbar";
+import { useCreateMeterFromName } from "./use-create-meter";
 import {
    buildBenefitColumns,
    type BenefitRow,
@@ -46,7 +43,6 @@ import {
 
 export function ServiceBenefitsTab({ serviceId }: { serviceId: string }) {
    const { openAlertDialog } = useAlertDialog();
-   const queryClient = useQueryClient();
    const [popoverOpen, setPopoverOpen] = useState(false);
    const [draftActive, setDraftActive] = useState(false);
 
@@ -66,14 +62,6 @@ export function ServiceBenefitsTab({ serviceId }: { serviceId: string }) {
       [meters],
    );
 
-   const linkedKey = useMemo(
-      () =>
-         orpc.benefits.getServiceBenefits.queryKey({
-            input: { serviceId },
-         }),
-      [serviceId],
-   );
-
    const attachedIds = useMemo(
       () => new Set(linked.map((b) => b.id)),
       [linked],
@@ -88,24 +76,7 @@ export function ServiceBenefitsTab({ serviceId }: { serviceId: string }) {
 
    const updateMutation = useMutation(
       orpc.benefits.updateBenefitById.mutationOptions({
-         meta: { skipGlobalInvalidation: true },
-         onMutate: async (vars) => {
-            await queryClient.cancelQueries({ queryKey: linkedKey });
-            const prev = queryClient.getQueryData<BenefitRow[]>(linkedKey);
-            if (prev) {
-               queryClient.setQueryData<BenefitRow[]>(
-                  linkedKey,
-                  prev.map((b) => (b.id === vars.id ? { ...b, ...vars } : b)),
-               );
-            }
-            return { prev };
-         },
-         onError: (e, _v, ctx) => {
-            if (ctx?.prev) queryClient.setQueryData(linkedKey, ctx.prev);
-            toast.error(e.message);
-         },
-         onSettled: () =>
-            queryClient.invalidateQueries({ queryKey: linkedKey }),
+         onError: (e) => toast.error(e.message),
       }),
    );
 
@@ -119,34 +90,7 @@ export function ServiceBenefitsTab({ serviceId }: { serviceId: string }) {
       }),
    );
 
-   const createMeterMutation = useMutation(
-      orpc.meters.createMeter.mutationOptions({
-         onError: (e) => toast.error(e.message),
-      }),
-   );
-
-   const handleCreateMeter = useCallback(
-      async (name: string) => {
-         const slug = name
-            .toLowerCase()
-            .normalize("NFD")
-            .replace(/[̀-ͯ]/g, "")
-            .replace(/[^a-z0-9]+/g, "_")
-            .replace(/^_|_$/g, "");
-         const created = await createMeterMutation.mutateAsync({
-            name,
-            eventName: slug || "custom_event",
-            aggregation: "sum",
-            filters: {},
-         });
-         await queryClient.invalidateQueries({
-            queryKey: orpc.meters.getMeters.queryKey({}),
-         });
-         toast.success(`Medidor "${name}" criado.`);
-         return created.id;
-      },
-      [createMeterMutation, queryClient],
-   );
+   const handleCreateMeter = useCreateMeterFromName();
 
    const handleSaveCell = useCallback(
       async (
