+++ apps/web/src/routes/_authenticated/$slug/$teamSlug/_dashboard/services/-services/service-prices-tab.tsx
@@ -7,11 +7,7 @@ import {
    EmptyMedia,
    EmptyTitle,
 } from "@packages/ui/components/empty";
-import {
-   useMutation,
-   useQueryClient,
-   useSuspenseQueries,
-} from "@tanstack/react-query";
+import { useMutation, useSuspenseQueries } from "@tanstack/react-query";
 import { CircleDollarSign, Copy, Plus, Trash2 } from "lucide-react";
 import { useCallback, useMemo, useState } from "react";
 import { toast } from "sonner";
@@ -30,12 +26,7 @@ import { ServiceTabToolbar } from "./service-tab-toolbar";
 
 export function ServicePricesTab({ serviceId }: { serviceId: string }) {
    const { openAlertDialog } = useAlertDialog();
-   const queryClient = useQueryClient();
    const [isDraftActive, setIsDraftActive] = useState(false);
-
-   const variantsKey = orpc.services.getVariants.queryKey({
-      input: { serviceId },
-   });
 
    const [{ data: prices }, { data: meters }] = useSuspenseQueries({
       queries: [
@@ -46,24 +37,7 @@ export function ServicePricesTab({ serviceId }: { serviceId: string }) {
 
    const updateMutation = useMutation(
       orpc.services.updateVariant.mutationOptions({
-         meta: { skipGlobalInvalidation: true },
-         onMutate: async (vars) => {
-            await queryClient.cancelQueries({ queryKey: variantsKey });
-            const prev = queryClient.getQueryData<ServicePrice[]>(variantsKey);
-            if (prev) {
-               queryClient.setQueryData<ServicePrice[]>(
-                  variantsKey,
-                  prev.map((p) => (p.id === vars.id ? { ...p, ...vars } : p)),
-               );
-            }
-            return { prev };
-         },
-         onError: (e, _v, ctx) => {
-            if (ctx?.prev) queryClient.setQueryData(variantsKey, ctx.prev);
-            toast.error(e.message);
-         },
-         onSettled: () =>
-            queryClient.invalidateQueries({ queryKey: variantsKey }),
+         onError: (e) => toast.error(e.message),
       }),
    );
